add validation

+++ backend/controllers/departmentController.js
@@ -1,4 +1,8 @@
+import mongoose from 'mongoose';
+
 import Department from '../models/departmentModel.js';
+import { validateDepartment } from '../validation/validateDepartment.js';
+
 
 export const getAllDepartments = async (req, res) => {
   try {
@@ -10,6 +14,11 @@ export const getAllDepartments = async (req, res) => {
 };
 
 export const getDepartment = async (req, res) => {
+
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(400).json({ message: 'Invalid Employee ID' });
+  }
+
   try {
     const department = await Department.findById(req.params.id);
     if (department == null) {
@@ -23,9 +32,16 @@ export const getDepartment = async (req, res) => {
 
 export const createDepartment = async (req, res) => {
   
-  // Validation for name,
-  if (!req.body.name) {
-    return res.status(400).json({ message: 'Name is required' });
+  // Validate the fields format and content
+  const { errors } = validateDepartment(req.body);
+  if (errors.length > 0) {
+    return res.status(400).json({ errors });
+  }
+
+  // Check if a department with the same name already exists
+  const existingDepartment = await Department.findOne({ name: req.body.name });
+  if (existingDepartment) {
+    return res.status(400).json({ message: 'A department with the same name already exists' });
   }
 
   const department = new Department({
@@ -35,6 +51,13 @@ export const createDepartment = async (req, res) => {
   });
   
   try {
+     // Validate the department before saving
+     const validationError = department.validateSync();
+     if (validationError) {
+       // If there's a validation error, send a 400 Bad Request response
+       res.status(400).json({ message: validationError.message });
+       return;
+     }
     const newDepartment = await department.save();
     res.status(201).json(newDepartment);
   } catch (err) {
@@ -43,6 +66,12 @@ export const createDepartment = async (req, res) => {
 };
 
 export const updateDepartment = async (req, res) => {
+    // Validate the fields format and content
+    const { errors } = validateDepartment(req.body);
+    if (errors.length > 0) {
+      return res.status(400).json({ errors });
+    }
+
   try {
     const updatedDepartment = await Department.findByIdAndUpdate(req.params.id, req.body, {
       new: true
@@ -57,6 +86,11 @@ export const updateDepartment = async (req, res) => {
 };
 
 export const deleteDepartment = async (req, res) => {
+
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(400).json({ message: 'Invalid Employee ID' });
+  }
+
     try {
       const deletedDepartment = await Department.findByIdAndDelete(req.params.id);
       if (deletedDepartment == null) {
