+++ backend/controllers/employeeController.js
@@ -1,6 +1,7 @@
-import Employee from '../models/employeeModel.js';
-import validator from 'validator';
 import mongoose from 'mongoose';
+
+import Employee from '../models/employeeModel.js';
+import { validateEmployee } from '../validation/validateEmployee.js'
 
 export const getAllEmployees = async (req, res) => {
   try {
@@ -12,6 +13,11 @@ export const getAllEmployees = async (req, res) => {
 };
 
 export const getEmployee = async (req, res) => {
+  
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(400).json({ message: 'Invalid Employee ID' });
+  }
+
   try {
     const employee = await Employee.findById(req.params.id);
     if (employee == null) {
@@ -58,99 +64,42 @@ export const createEmployee = async (req, res) => {
       res.status(400).json({ message: `Missing ${missingFields.join(', ')} in request body` });
       return;
     }
-
-    const {
-      firstName,
-      lastName,
-      email,
-      mobilePhone,
-      hireDate,
-      department,
-      jobTitle,
-      nationalIDNumber,
-      socialSecurityNumber,
-      nationality,
-      gender,
-      address,
-      photo,
-      bloodType,
-      martialStatus,
-      militaryStatus,
-      educationStatus
-    } = req.body;
-
-  // specific validations
-  if (!validator.isEmail(email)) {
-    return res.status(400).json({ message: 'Invalid Email Address' });
+ 
+  // Validate the fields format and content  
+  const { isValid, errors } = validateEmployee(req.body);
+  if (!isValid) {
+    return res.status(400).json({ errors });
   }
 
-  if (!validator.isMobilePhone(mobilePhone)) {
-    return res.status(400).json({ message: 'Invalid Mobile Phone Number' });
-  }
-  // mobile phone of a relative in case of
-  const homePhone = req.body.homePhone;
-  if (homePhone && !validator.isMobilePhone(homePhone)) {
-    return res.status(400).json({ message: 'Invalid Emergency Contact Number' });
-  }
-
-  const allowedGenders = ['Male', 'Female', 'Other'];
-  if (!allowedGenders.includes(gender)) {
-    return res.status(400).json({ message: 'Invalid Gender' });
-  }
-
-  const allowedBloodTypes = ['A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-'];
-  if (!allowedBloodTypes.includes(bloodType)) {
-    return res.status(400).json({ message: 'Invalid Blood Type' });
-  }
-
-  const allowedMartialStatuses = ['Single', 'Married', 'Divorced', 'Widowed'];
-  if (!allowedMartialStatuses.includes(martialStatus)) {
-    return res.status(400).json({ message: 'Invalid Martial Status' });
-  }
-
-  const allowedMilitaryStatuses = ['Active', 'Reserve', 'Veteran', 'None'];
-  if (!allowedMilitaryStatuses.includes(militaryStatus)) {
-    return res.status(400).json({ message: 'Invalid Military Status' });
-  }
-
-  const allowedEducationStatuses = ['High School', 'Associate', 'Bachelor', 'Master', 'Doctorate'];
-  if (!allowedEducationStatuses.includes(educationStatus)) {
-    return res.status(400).json({ message: 'Invalid Education Status' });
-  }
-
-  const validHireDate = new Date(hireDate);
-  if (!validHireDate.getTime()) {
-    return res.status(400).json({ message: 'Invalid Hire Date' });
-  }
-
-  if (!mongoose.Types.ObjectId.isValid(department)) {
-    return res.status(400).json({ message: 'Invalid Department ID' });
-  }
-
-  if (validator.isEmpty(photo)) {
-    return res.status(400).json({ message: 'Photo is required' });
-  }
+   // Check if an employee with the same email address already exists
+   const existingEmployee = await Employee.findOne({ email: req.body.email });
+   if (existingEmployee) {
+     return res.status(400).json({ message: 'An employee with the same email address already exists' });
+   } 
 
 
-  const employee = new Employee({
-    firstName,
-    lastName,
-    email,
-    mobilePhone,
+   const employee = new Employee({
+    firstName: req.body.firstName,
+    lastName: req.body.lastName,
+    email: req.body.email,
+    mobilePhone: req.body.mobilePhone,
     homePhone: req.body.homePhone,
-    hireDate,
-    department,
-    jobTitle,
-    nationalIDNumber,
-    socialSecurityNumber,
-    nationality,
-    gender,
-    address,
-    photo,
-    bloodType,
-    martialStatus,
-    militaryStatus,
-    educationStatus
+    hireDate: new Date(req.body.hireDate),
+    department: req.body.department,
+    jobTitle: req.body.jobTitle,
+    nationalIDNumber: req.body.nationalIDNumber,
+    socialSecurityNumber: req.body.socialSecurityNumber,
+    nationality: req.body.nationality,
+    gender: req.body.gender,
+    address: req.body.address,
+    photo: req.body.photo,
+    bloodType: req.body.bloodType,
+    martialStatus: req.body.martialStatus,
+    militaryStatus: req.body.militaryStatus,
+    educationStatus: req.body.educationStatus,
+    emergencyContactName: req.body.emergencyContactName,
+    emergencyContactRelationship: req.body.emergencyContactRelationship,
+    emergencyContactPhone: req.body.emergencyContactPhone
   });
 
   try {
@@ -171,20 +120,33 @@ export const createEmployee = async (req, res) => {
 };
 
 export const updateEmployee = async (req, res) => {
-    try {
-      const updatedEmployee = await Employee.findByIdAndUpdate(req.params.id, req.body, {
-        new: true
-      });
-      if (updatedEmployee == null) {
-        return res.status(404).json({ message: 'Cannot find employee' });
-      }
-      res.json(updatedEmployee);
-    } catch (err) {
-      res.status(400).json({ message: err.message });
+  // Validate the fields format and content
+  const { errors } = validateEmployee(req.body);
+  if (errors.length > 0) {
+    res.status(400).json({ errors });
+    return;
+  }
+
+  try {
+    const updatedEmployee = await Employee.findByIdAndUpdate(req.params.id, req.body, {
+      new: true
+    });
+    if (updatedEmployee == null) {
+      return res.status(404).json({ message: 'Cannot find employee' });
     }
-  };
+    res.json(updatedEmployee);
+  } catch (err) {
+    res.status(400).json({ message: err.message });
+  }
+};
+
 
 export const deleteEmployee = async (req, res) => {
+
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(400).json({ message: 'Invalid Employee ID' });
+  }
+
     try {
       const deletedEmployee = await Employee.findByIdAndDelete(req.params.id);
       if (deletedEmployee == null) {
